Added table-numbers
1.0
Added table numbers which is a solver for a Math Olympiad problem. It can find all of the sequences of the numbers, which must together give a number as a result and the current highest number must be divisible by the sum of all other numbers combined.
1.1
added anchor to the main page
1.2
added description of the solver